backend/requirements.txt
Flask==3.1.3
Flask-Limiter==4.1.1
Flask-Migrate==4.1.0
flask-cors==6.0.2
flask-jwt-extended==4.7.1
SQLAlchemy==2.0.41
alembic==1.17.0
celery==5.5.2
redis==6.2.0
celery-redbeat==2.2.0
gunicorn==23.0.0
pydantic==2.11.5
werkzeug==3.1.3
marshmallow==3.26.1
psycopg2-binary==2.9.10
structlog==25.1.0
python-dotenv==1.0.1
requests==2.32.3
h3==4.2.1
numpy==2.1.3
scikit-learn==1.6.1
geoalchemy2==0.17.1
httpx==0.28.1
geopy==2.4.1
overpy==0.7
sparqlwrapper==2.0.0
wikipedia-api==0.8.1
Pillow==11.1.0
tiktoken==0.9.0
pip-audit==2.8.0
